tests/wpt/tests/css/css-view-transitions/scoped/view-transition-is-anchor-auto.html

<!DOCTYPE html>
<html>
<head>
  <link rel="help" href="https://www.w3.org/TR/css-view-transitions-2/">
  <meta name="author" content="vmpstr@chromium.org">
  <title>Position anchor is auto on view transition</title>
</head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="scope" class="block"></div>

<script>
promise_test(() => {
  return scope.startViewTransition().ready.then(() => {
    assert_equals(getComputedStyle(scope, "::view-transition").positionAnchor, "auto");
  });
});
</script>
</html>
